git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@5076 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/legacy-test-tree/tests.base/com/tc/test/activepassive/ActivePassiveServerManager.java
@@ -611,7 +611,7 @@ public class ActivePassiveServerManager {
       crashRandomServer();
     }
 
-    if (serverPersistence.equals(ActivePassivePersistenceMode.PERMANENT_STORE)) {
+    if (serverNetworkShare && serverPersistence.equals(ActivePassivePersistenceMode.PERMANENT_STORE)) {
       System.out.println("Deleting data directory for server=[" + servers[lastCrashedIndex].getDsoPort() + "]");
       deleteDirectory(serverConfigCreator.getDataLocation(lastCrashedIndex));
     }
